chore: navigation added to project to allow scalability with fragments
