+++ app/src/main/java/com/picpay/desafio/android/presentation/home/UserListDiffCallback.kt
@@ -1,4 +1,4 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.presentation.home
 
 import androidx.recyclerview.widget.DiffUtil
 import com.picpay.desafio.android.domain.entity.User
