+++ app/src/main/java/com/picpay/desafio/android/presentation/home/UserListItemViewHolder.kt
@@ -1,7 +1,8 @@
-package com.picpay.desafio.android
+package com.picpay.desafio.android.presentation.home
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
+import com.picpay.desafio.android.R
 import com.picpay.desafio.android.domain.entity.User
 import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
