+++ app/src/main/res/navigation/agenda_nav_graph.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/agenda_nav_graph.xml"
+    app:startDestination="@id/agendaFragment">
+
+    <fragment
+        android:id="@+id/agendaFragment"
+        android:name="com.picpay.desafio.android.presentation.home.AgendaFragment"
+        android:label="AgendaFragment"
+        tools:layout="@layout/fragment_agenda"/>
+
+</navigation>
