+++ build.gradle
@@ -9,6 +9,7 @@ buildscript {
         core_testing_version = '2.1.0'
         constraintlayout_version = '1.1.3'
         material_version = "1.1.0"
+        navigation_version = "2.3.5"
         moshi_version = '1.8.0'
         retrofit_version = '2.9.0'
         okhttp_version = '4.3.1'
